VBP- python files/requirements.txt
PyAudio==0.2.11
SpeechRecognition==3.8.1
inflect==4.1.0
libindic.soundex==1.0.2
libindic.utils==1.0.3
word2number==1.1
pocketsphinx==0.1.15
fuzzy==1.2.2